Finished fc.gate.density. Cleaned up other issues.
Finished implementing fc.gate.density. Tried a few different implementations,
but the straight up MATLAB port seemed to run the fastest, so I went with
that.

Also reorganized some plotting code to acknowledge that multi-dimensional
numpy arrays need to be manipulated to appear as we expect in a plot.

Added labels to colorbars for fc.plot.fsc_ssc_counts and
fc.plot.fsc_ssc_density.