Removed check on number of channels in FCS file
The CHANNEL_INFO dictionaries are populated from the TEXT dictionary via
the get() method which returns None if a key is absent. This gracefully
fails in the case where less than 6 channels exist in a FCS file, so I
just decided to remove the check on number of channels.

Fixes #1