Updated extraction of detector voltage by FCSData object.
Updated extraction of `detector_voltage` to mirror extraction of
`amplifier_gain`. Specifically, this means `FCSData` will attempt
to extract `detector_voltage` from the "$PnV" keyword parameter
first, and only if that value is `None` will it go on to check and
see if the machine is a CellQuest Pro machine (in which case it
will attempt to extract `detector_voltage` from an alternative set
of keyword parameters).

See #227.