Bug fix for io.read_fcs_text_segment(). More unit tests.
Fixed bug that occurs when parsing TEXT segments where keywords or
keyword values have delimiters. Added a bunch of unit tests to test
a bunch of edge cases related to this.

Removed some previous unit tests because they didn't test an explicit
violation of the FCS standards or an edge case.

See #247.