Fix reading .xls files with new pandas/openpyxl.
When reading an Excel file with pandas using the openpyxl engine,
pandas >= v1.2.0 opens the file and passes the file buffer to
openpyxl, which may determine the file is not a zip file (e.g., if
it's a .xls file) and throw a previously unhandled
zipfile.BadZipFile exception. This fix catches this new exception
and tries to read the file with the xlrd engine. (pandas < v1.2.0
passed the file name to openpyxl, which inspected its extension and
raised an openpyxl.utils.exceptions.InvalidFileException. FlowCal
still looks for this exception as well.)